pyproject.toml
[project]
name = "adk-tutorials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.21.0",
    "litellm>=1.80.10",
    "ollama>=0.6.1",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.0",
    "yfinance>=0.2.66",
]
